Fix conversion of numpy arrays of higher rank to array<array<..>>
Numpy will convert an array of arrays to a an array of higher rank
if possible
   array([array([..]), ..]) -> array([[..], ..])
This has to be accounted for in both the py2c and is_convertible of nda::array.
We cannot just call is_convertible / py2c on each element of the higher-rank numpy array.
Instead we must access into it using __getitem__ with only R<=r indices,
where R is the rank of our nda::array and r the rank of the numpy array.