Add float32 support for some lapack routines
This was a bit hairier than the blas patch before, as there were a lot of implicit assumptions
about a double underlying type in many of the temporaries and null returns.  For tests I
factored out some logic to make the tests run in float and double, which also required grabbing
separate tolerances and passing them to the checkers. I've added some more tests to teh blas
and linear algebra test runners. Also, one of the eigenvector/value decompositions gives
eigenvectors with a different sign, which is something that probably needs to be handled better
than my approach. This isn't wrong, and there are means to ensure the same eigenvector sign, if
that's of interest in the future.

Some other functions were touched, but these are the main functions in this patch:
geqp3, heev, orgqr, syev, ungqr, eigh